--- check/find_subjects.go
@@ -1,4 +1,0 @@
-package check
-
-type SubjectFinder struct {
-}
